Added the IT subdirectory with treatment of W-types in HoTT.
The IT directory, which was previously part of the HoTT subdirectory,
properly belongs to the Foundations repository, as it depends on it.
Thus we have moved it over here, also to maintain compatibility with
future versions of Foundations.

The IT subdirectory does not compile yet, however. While I have
fixed Makefile, I think, there seem to be a number of
"Add Rec Load Path", which need to be modified, or even better,
they should be removed because they are already there in the
Makefile so they don't surve a purpose anymore. But I think it
is not much work to get IT to compile.